fix: move ordering test to sibling file (anti-monolith ceiling)
Extract reconcile_runs_before_fetch_loop_ordering test from
worktree_cleanup.rs (2617 LOC) into sibling
worktree_cleanup/reconcile_ordering_tests.rs.

worktree_cleanup.rs: 2475 LOC (under 2500 ceiling).


Agend-Agent: claude-aef7c0
Agend-Task: t-20260717085016945919-2031-1
Agend-Branch: fix/legacy-reconcile-before-scm-sweep
Agend-Issued-At: 2026-07-17T08:50:31.820549+00:00
